Extract response fields from response headers
Extract values from response headers based in rules specified in the
`additional_bindings` section of an `HttpRule`, using the `custom`
pattern with a `kind` of `response_header`. This follows a similar
pattern spec to request headers, but the spec is used to parse response
headers rather than generate request headers.

Only singular or repeated scalar fields can be extracted, and only into
the top level of the response proto. A pattern can be used to extract a
single header to multiple fields and also require that literal text that
is not extracted also matches.

Headers are extracted after the response body is unmarshaled into the
response, so that headers can override the body.

If a header appears multiple times, each is extracted. If any fields in
the target message are not repeated, the value parsed from the last
header will be the value in the target message. If the field(s) are
repeated fields, they are all extracted in the order that the headers
appear.