httprule: Add headers to ClientConn
Add a constructor for ClientConn that takes Options, and add an Option
to set a HTTP Header on each HTTP request it makes. This allows a user
of the ClientConn to set authorization headers and such where those
headers are not specified in the `.proto` file but need to be set more
dynamically.

Also add an Option for overriding the default HTTPClient used by the
ClientConn.