Transplant registry
Transplant registry from jig.
Use protosync to download google api files.

registry.Files is wrapper for the official protoregistry.Files,
implementing protoregistry.MessageTypeResolver and
protoregistry.ExtensionTypeResolver, so that we can easily use this
Files wrapper with prototest and protojson to unmarshal extension.

A registry is created from a FileDescriptorSet like so:

	var fds *descriptorpb.FileDescriptorSet              // initialise
	protoregistryFiles, err := protodesc.NewFiles(&fds)  // handle err
	files := registry.NewFiles(protoregistryFiles)

link: https://github.com/foxygoat/jig/tree/v0.0.25/registry
Signed-off-by: Julia Ogris <julia.ogris@gmail.com>