pb: Add --descriptorpb/-D flag for easy decoding of descriptor pbs
Add a `--descriptorpb`/`-D` flag that specifies that descriptorpb should
be used as the `FileDescriptorSet` instead of requiring it to be
specified on the command line. This makes `--protoset`/`-P` not required
in this case.

This makes it easy to inspect what `protoc` produces:

    protoc -o/dev/stdout foo.proto | pb -D

This will show the descriptorpb form of `foo.proto` in JSON (or add
`-Ot` to see it in prototext format, which is a bit more compact).