registry: Populate Types with dynamic messages
Add some helper functions to populate a `protoregistry.Types` registry
with `dynamicpb.Message` values created from a `FileDescriptorSet`.

This is an alternative to the `Files` registry in this package that
implements the *Resolver interfaces creating dynamic messages on demand.

The benefit of doing it this way is that existing concrete types are
retained in the Types registry so it is possible to type assert a
message created from the Types registry to a concrete type compiled into
the binary. When you have some known types (such as the well-known proto
types as well as descriptorpb), it is easier to manipulate these using
the concrete type instead of using the reflection-like interface of
`dynamicpb.Message`.

One helper function (`CloneTypes()`) clones a `protoregistry.Types`
registry so you can start with the `protoregistry.GlobalTypes` registry
of built-in types and then use the second helper (`AddDynamicTypes()`)
to add dynamic types from a `FileDescriptorSet` to that registry.