pb: Decode input FileDescriptorSets with self in registry
When decoding input of message type `FileDescriptorSet`, add that input
to the registry used for resolving extensions and messages. This allows
option extensions defined in the input to be used in options in the
`FileDescriptorSet`. This requires double-unmarshaling of the input as
it needs to be unmarshaled first to add to the registry, and again to
unmarshal with that registry. The first unmarshal does not decode the
extensions defined in the input, so the second is necessary.

Add some tests for this with a new proto file and a golden file to match
the output against.