proto: Regenerate/sync protos
Run

    make proto

to re-sync and regenerate protos. It looks like FileDescriptorProto in
descriptorpb has a new field `edition`, which has caused regtest.pb to
change (it includes imports, which is descriptorpb). The new edition
field (and `editions` syntax) is completely undocumented as to its
meaning. Thanks Google.

Update the registry tests as they depend on the number of types defined
in descriptor.proto and descriptor.proto has added a new enum. I should
make these tests independent of third-party protos, but these ones
contain a good bunch of different types that it pushes the registry a
bit harder. It doesn't change all that often.