compiler: Add participle-based protobuf parser
Add a protobuf parser using participle, based on the example in Alec
Thomas's protobuf compiler. I have reworked the parser to be more how I
think it should look and to bring it closer to the BNF(ish) in the
protobuf spec. I have also removed comments from the grammar (they are
ignored like whitespace) since I don't believe the best way to handle
comments will be in the grammar (I am thinking of attaching comments to
tokens in the lexer; we'll see).

At the moment, the `protog` command just prints out the Go struct for
the parsed form of the protobuf. This will grow into something more
useful as the compiler evolves to be a proper compiler.

We bump the required Go version to 1.18 as there is a small use of
generics in the test code, but I also want to use generics wherever it
makes sense.

Link: https://github.com/alecthomas/participle
Link: https://github.com/alecthomas/protobuf
Link: https://developers.google.com/protocol-buffers/docs/reference/proto2-spec
Link: https://developers.google.com/protocol-buffers/docs/reference/proto3-spec