build: Ignore staged changes on up-to-date check
Have `make check-uptodate` ignore staged changes so you can run `make
ci` with all your changes staged to verify all is good before
committing. The `check-uptodate` target leaves files dirty in the
workspace so this change should not alter the original desired
behaviour.

`git status --porcelain` outputs a non-space character as the first
character on a line for staged changes, unless it is `?` which means
untracked.